refactor: extract routers/capture.py
Move /ui/{ctx}/capture* routes out of main.py into api/routers/capture.py.
Wire router into main.py.
